projects/chapter-12-project/requirements.txt
# Chapter 12 Project Requirements

# Core AI library
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0
